Migrate kernel builder in sfh.py into kernel_builders.py
