+++ diffstar/tests/test_lax_sfh.py
@@ -4,7 +4,7 @@ import numpy as np
 from jax import jit as jjit
 from jax import vmap
 
-from ..sfh import get_sfh_from_mah_kern
+from ..kernel_builders import get_sfh_from_mah_kern
 from .test_diffstar_is_frozen import (
     _get_default_mah_params,
     _get_default_sfr_u_params,
